Conflicts:
	Cargo.lock
	dapps/src/rpc.rs

+++ util/bigint/src/hash.rs
@@ -25,7 +25,7 @@ use std::str::FromStr;
 use rand::Rng;
 use rand::os::OsRng;
 use rustc_serialize::hex::{FromHex, FromHexError};
-use uint::{Uint, U256};
+use bigint::{Uint, U256};
 
 /// Trait for a fixed-size byte array to be used as the output of hash functions.
 pub trait FixedHash: Sized {
@@ -512,7 +512,7 @@ pub type H256FastSet = HashSet<H256, BuildHasherDefault<PlainHasher>>;
 #[cfg(test)]
 mod tests {
 	use hash::*;
-	use uint::*;
+	use bigint::*;
 	use std::str::FromStr;
 
 	#[test]
